+++ tests/testthat/test-select.r
@@ -124,6 +124,19 @@ test_that("rename handles grouped data (#640)", {
   expect_equal(names(res), c("a", "c"))
 })
 
+test_that("rename does not crash with invalid grouped data frame (#640)", {
+  df <- data_frame(a = 1:3, b = 2:4, d = 3:5) %>% group_by(a, b)
+  df$a <- NULL
+  expect_equal(
+    df %>% rename(e = d) %>% ungroup,
+    data_frame(b = 2:4, e = 3:5)
+  )
+  expect_equal(
+    df %>% rename(e = b) %>% ungroup,
+    data_frame(e = 2:4, d = 3:5)
+  )
+})
+
 # combine_vars ------------------------------------------------------------
 # This is the low C++ function which works on integer indices