+++ NEWS.md
@@ -22,6 +22,9 @@
 
 * `summarise()` now can create ordered factors (#2200).
 
+* Fixed segmentation fault after calling `rename()` on an invalid grouped
+  data frame (#2031).
+
 * Breaking change: `xxx_join.tbl_df()` by default treats all `NA` values as
   different from each other (and from any other value), so that they never
   match.  This corresponds to the behavior of joins for database sources,